Switch class names
